refactor:#21 InputForm 공통컴포넌트로 변경
